chore: changes datatypes

+++ prisma/migrations/20230921005158_initial_setup/migration.sql
@@ -2,6 +2,6 @@
 CREATE TABLE "points_of_interest" (
     "id" TEXT NOT NULL PRIMARY KEY,
     "name" TEXT NOT NULL,
-    "x_coord" BIGINT NOT NULL,
-    "y_coord" BIGINT NOT NULL
+    "x_coord" INTEGER NOT NULL,
+    "y_coord" INTEGER NOT NULL
 );
